Made a base class for Leumi and Massad as requested in the PR
